moved enums to their own file, moved generate_documentation_background function to DocumentationService. Added the status code 202 for POST.
