+++ routers/docs.py
@@ -3,39 +3,17 @@ import logging
 
 from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException
 
-from schemas.api import GenerateFileDocsRequest, GenerateFileDocsResponse, GetFileDocsResponse, DocsStatusEnum
+from schemas.api import GenerateFileDocsRequest, GenerateFileDocsResponse, GetFileDocsResponse
 from schemas.database import FirestoreDocumentationCreateModel, FirestoreDocumentationUpdateModel
+from schemas.documentation_generation_enums import DocsStatusEnum
 
 from services.documentation_service import DocumentationService, get_documentation_service
 from services.clients.firebase_client import FirebaseClient, get_firebase_client
 
+
 router = APIRouter()
 
-# background tasks for generating the documentation
-async def generate_documentation(
-    documentation_service: DocumentationService,
-    firebase_client: FirebaseClient,
-    firebase_file_id: str,
-    github_url: str, 
-):
-    generated_docs = await documentation_service.generate_doc_for_github_file(github_url)
-
-    blob_url = firebase_client.get_blob_url(firebase_file_id)
-
-    firebase_client.add_blob(blob_url, generated_docs.content)
-    
-    firebase_client.update_document(
-        FirebaseClient.TEST_COLLECTION, 
-        firebase_file_id, 
-        FirestoreDocumentationUpdateModel
-        (
-            bucket_url=blob_url,
-            status=DocsStatusEnum.COMPLETED
-        ).model_dump()
-    )
-
-
-@router.post("/file-docs")
+@router.post("/file-docs", status_code=http.client.ACCEPTED)
 async def generate_file_docs(
         generate_file_docs_request: GenerateFileDocsRequest,
         background_tasks: BackgroundTasks,
@@ -56,7 +34,12 @@ async def generate_file_docs(
         )
 
         # add task to be done async
-        background_tasks.add_task(generate_documentation, documentation_service, firebase_client, document_ref.id, generate_file_docs_request.github_url)
+        background_tasks.add_task(
+            documentation_service.generate_documentation_background, 
+            firebase_client, 
+            document_ref.id, 
+            generate_file_docs_request.github_url
+        )
 
         return GenerateFileDocsResponse(
             message="Documentation generation has been started.",
