+++ schemas/database.py
@@ -1,5 +1,5 @@
 from pydantic import BaseModel
-from schemas.api import DocsStatusEnum
+from schemas.documentation_generation_enums import DocsStatusEnum
 
 class FirestoreDocumentationCreateModel(BaseModel):
     bucket_url: str | None
